requirements.txt
Flask
pillow
gunicorn
tensorflow
numpy
request